spore postepy

+++ lab1/zad4_klient.sh
@@ -12,10 +12,15 @@ serwer_pid=`pidof zad4_serwer.sh`
 
 wait $serwer_pid
 
+if [ -s zad4_wyniki.txt ]
+then
+
 wynik=$(cat zad4_wyniki.txt)
 
 echo $wynik
 
 echo > zad4_dane.txt
+
+fi
 
 exit
